Issue #31/#36: Revisit sound modes for different channels
Phone mute/vibrate setting only applies to notification channel

Previously the global phone mute/vibrate notification setting
was applied to notification sound regardless of the selected
notification channel. This kind of defeats the purpose of having
separate notification channels.

This commit changes the behaviour to only consider the phone
mute/vibrate setting if the 'notification' channel is used.
If the sound channel is set to 'media' or 'alarm' the mindful
bell notifications will sound regardless of the phone mute/vibrate
setting. In this case you can still mute the bell via the
app 'Mute' toggle switch on the main app screen.

Summary:
- App mute is enabled: always mute
- For notification channel: honour the phone silent/vibrate-only setting
- For media and alarm channels, allow sound unless DND is enabled